refactor(intake): close intake parity checkpoint
Why:
- replace family-level parity tracking with a mechanically checked behavior ledger
- finish the intake area while its routing, file-fact, and archive seams are open

What:
- retire the duplicate parity matrix and add a master-ledger validator
- split intake archive, file-facts, and routing logic into nested packages with mirrored tests
- update parity docs and engineering standards to the final intake structure

Checks:
- uv run pytest tests/unit/test_master_parity_ledger.py tests/unit/test_docs_runtime_parity.py tests/unit/application/services/intake --no-cov -q
- uv run pre-commit run markdownlint --all-files
- uv run python -m tools.run_quality_gates --full-tests
