+++ src/crypto_reconciliation/application/services/intake/file_facts/inspection.py
@@ -1,10 +1,9 @@
-"""Typed file facts used by intake routing and review decisions."""
+"""CSV inspection and signal extraction for intake files."""
 
 from __future__ import annotations
 
 import csv
 import re
-from dataclasses import dataclass
 from datetime import UTC, datetime
 from pathlib import Path
 
@@ -15,6 +14,8 @@ from crypto_reconciliation.domain.wallet_identifiers import (
     TRON_ADDRESS_PATTERN,
     scope_token_for_identifier,
 )
+
+from .models import IntakeFileFacts
 
 type CsvCell = str | list[str]
 
@@ -49,25 +50,6 @@ NETWORK_HINTS = (
     ("solana", "solana"),
 )
 ACCOUNT_SEGMENT_PATTERN = re.compile(r"account[-_ ]?[a-z0-9]+", re.IGNORECASE)
-PATH_DATE_PATTERNS = (
-    re.compile(r"(?<!\d)(?P<year>20\d{2})[-_.](?P<month>\d{2})[-_.](?P<day>\d{2})(?!\d)"),
-    re.compile(r"(?<!\d)(?P<month>\d{2})[-_.](?P<day>\d{2})[-_.](?P<year>20\d{2})(?!\d)"),
-    re.compile(r"(?<!\d)(?P<year>20\d{2})(?P<month>\d{2})(?P<day>\d{2})(?:\d{4})?(?!\d)"),
-)
-PATH_MONTH_PATTERNS = (
-    re.compile(r"(?<!\d)(?P<year>20\d{2})[-_/](?P<month>\d{2})(?!\d)"),
-    re.compile(r"(?<!\d)(?P<year>20\d{2})(?P<month>\d{2})(?!\d)"),
-)
-PATH_YEAR_PATTERN = re.compile(r"(?<!\d)(20\d{2})(?!\d)")
-
-
-@dataclass(frozen=True)
-class IntakeFileFacts:
-    header: tuple[str, ...] = ()
-    min_timestamp: str = ""
-    max_timestamp: str = ""
-    scope_tokens: tuple[str, ...] = ()
-    network_hints: tuple[str, ...] = ()
 
 
 def inspect_intake_file(path: Path, *, relative_path: str) -> IntakeFileFacts:
@@ -90,12 +72,15 @@ def inspect_intake_file(path: Path, *, relative_path: str) -> IntakeFileFacts:
     )
 
 
-def detect_capture_id(relative_path: str, facts: IntakeFileFacts) -> str:
-    if facts.min_timestamp:
-        parsed = _parse_timestamp(facts.min_timestamp)
-        if parsed is not None:
-            return parsed.strftime("%Y-%m")
-    return _capture_id_from_path(relative_path)
+def parse_timestamp(value: str) -> datetime | None:
+    if not value:
+        return None
+    for fmt in TIMESTAMP_FORMATS:
+        try:
+            return datetime.strptime(value, fmt).replace(tzinfo=UTC)
+        except ValueError:
+            continue
+    return None
 
 
 def _timestamp_values(rows: list[dict[str, CsvCell]], header: tuple[str, ...]) -> list[str]:
@@ -105,9 +90,7 @@ def _timestamp_values(rows: list[dict[str, CsvCell]], header: tuple[str, ...]) -
         field_name = next((lowered[name.lower()] for name in TIMESTAMP_FIELD_NAMES if name.lower() in lowered), "")
     if not field_name:
         return []
-    parsed_values = [
-        parsed for row in rows if (parsed := _parse_timestamp(_cell_text(row.get(field_name)))) is not None
-    ]
+    parsed_values = [parsed for row in rows if (parsed := parse_timestamp(_cell_text(row.get(field_name)))) is not None]
     return [value.strftime("%Y-%m-%d %H:%M:%S") for value in sorted(parsed_values)]
 
 
@@ -121,17 +104,6 @@ def _read_csv_rows(path: Path) -> tuple[tuple[str, ...], list[dict[str, CsvCell]
             dialect = csv.excel
         reader = csv.DictReader(handle, dialect=dialect)
         return tuple(reader.fieldnames or ()), list(reader)
-
-
-def _parse_timestamp(value: str) -> datetime | None:
-    if not value:
-        return None
-    for fmt in TIMESTAMP_FORMATS:
-        try:
-            return datetime.strptime(value, fmt).replace(tzinfo=UTC)
-        except ValueError:
-            continue
-    return None
 
 
 def _scope_tokens(relative_path: str, rows: list[dict[str, CsvCell]]) -> set[str]:
@@ -185,32 +157,3 @@ def _identifier_scope_tokens(text: str) -> set[str]:
             if token:
                 tokens.add(token)
     return tokens
-
-
-def _capture_id_from_path(relative_path: str) -> str:
-    parsed_dates = sorted(_parsed_path_dates(relative_path))
-    if parsed_dates:
-        return parsed_dates[-1].strftime("%Y-%m")
-    for pattern in PATH_MONTH_PATTERNS:
-        match = pattern.search(relative_path)
-        if match is not None:
-            return f"{match.group('year')}-{match.group('month')}"
-    year_match = PATH_YEAR_PATTERN.search(relative_path)
-    if year_match is not None:
-        return year_match.group(1)
-    return ""
-
-
-def _parsed_path_dates(relative_path: str) -> list[datetime]:
-    parsed_dates: list[datetime] = []
-    for pattern in PATH_DATE_PATTERNS:
-        for match in pattern.finditer(relative_path):
-            if "month" in match.groupdict() and "day" in match.groupdict() and "year" in match.groupdict():
-                month = int(match.group("month"))
-                day = int(match.group("day"))
-                year = int(match.group("year"))
-                try:
-                    parsed_dates.append(datetime(year, month, day, tzinfo=UTC))
-                except ValueError:
-                    continue
-    return parsed_dates
