+++ src/crypto_reconciliation/application/services/intake/routing/__init__.py
@@ -1,0 +1,7 @@
+"""Typed intake routing rules."""
+
+from .classification import detect_source_folder
+from .models import IntakeRoute
+from .service import route_intake_file
+
+__all__ = ["IntakeRoute", "detect_source_folder", "route_intake_file"]
